% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadran.R
\name{cadran}
\alias{cadran}
\title{draw a wath}
\usage{
cadran(t = Sys.time(), sec = TRUE, coord.h = circleFun(npoints = 60 *
  12, r = 0.8), coord.m = circleFun(npoints = 60 * 12),
  coord.s = coord.m)
}
\arguments{
\item{t}{time as character POSIXct or Date}

\item{sec}{booleen display needle second hand}

\item{coord.h}{coordinates of hours}

\item{coord.m}{coordinates of minutes}

\item{coord.s}{coordinates of seconds}

\item{color.h}{hour needle color}

\item{color.m}{minute needle color}

\item{color.s}{second needle color}
}
\description{
draw a wath
}
\examples{
cadran()
cadran(t="11:45:23")
cadran(t="07:46:23")
cadran(t="07:46:23",sec=FALSE)
cadran(t="07:46")

}
