% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_digital.R
\name{draw_as_digits}
\alias{draw_as_digits}
\title{draw time using digits number}
\usage{
draw_as_digits(t = lubridate::now(), size = 200, ampm = FALSE)
}
\arguments{
\item{t}{time to draw}

\item{size}{size in pixel}

\item{ampm}{if FALSE (by default) use 24H clock}
}
\description{
draw time using digits number
}
\examples{
draw_as_digits(lubridate::now())
}
