% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clock.R
\name{mod_clock_ui}
\alias{mod_clock_ui}
\alias{mod_clock_server}
\title{mod_clock_ui and mod_clock_server}
\usage{
mod_clock_ui(id)

mod_clock_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
