% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_digit.R
\name{mod_digit_ui}
\alias{mod_digit_ui}
\alias{mod_digit_server}
\title{mod_digit_ui and mod_digit_server}
\usage{
mod_digit_ui(id)

mod_digit_server(input, output, session, freq = 1000)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
